requirements.txt
Django==3.2.5
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
asgiref==3.4.1
python-decouple==3.3
PyJWT==2.1.0